requirements.txt
# Core
fastapi>=0.104.1,<0.110.0
python-dotenv>=1.0.0,<2.0.0
uvicorn>=0.24.0,<0.30.0

# Database
supabase>=2.0.3,<3.0.0

# Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0

# Validation
pydantic>=2.6.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
email-validator>=2.1.0,<3.0.0

# Development
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<0.24.0
httpx>=0.24.0,<0.28.0